Relpace System::LightFactory with unified LightFactory
