+++ lib/stoplight/wiring/light_factory.rb
@@ -2,15 +2,14 @@
 
 module Stoplight
   module Wiring
-    # Constructs a fully-wired Light instance from validated configuration.
-    #
-    # Wires together all infrastructure components (trackers, strategies) needed
-    # for a functioning circuit breaker. Subclasses supply the storage backend
-    # (state, metrics, recovery-lock stores).
+    # Wires together all infrastructure components (trackers, strategies, storage
+    # backend) needed for a functioning circuit breaker.
     #
     # @api private
     class LightFactory
-      def initialize(config:)
+      def initialize(system_name:, config:, failover_system:)
+        @system_name = system_name
+        @failover_system = failover_system
         @config = config
 
         @clock = Infrastructure::SystemClock.new
@@ -41,6 +40,9 @@ module Stoplight
         )
       end
 
+      def state_store = storage_set.state_store
+      def metrics_store = storage_set.metrics_store
+
       private
 
       attr_reader :data_store_config
@@ -49,6 +51,21 @@ module Stoplight
       attr_reader :traffic_control
       attr_reader :traffic_recovery
       attr_reader :config
+      attr_reader :system_name
+
+      def recovery_lock_store = storage_set.recovery_lock_store
+      def recovery_metrics_store = storage_set.recovery_metrics_store
+      def storage_scripting = Infrastructure::Redis::Storage::Scripting.new(redis:)
+      def failover_system = T.must(@failover_system)
+
+      def key_space = @key_space ||= Infrastructure::Redis::Storage::KeySpace.build(
+        system_name: system_name,
+        light_name: config.name
+      )
+
+      def storage_set
+        @storage_set ||= StorageSetBuilder.new(backend: build_backend, windowed: !config.window_size.nil?).build
+      end
 
       # @return [<Stoplight::Notifier::Base>]
       def notifiers
@@ -107,6 +124,22 @@ module Stoplight
         else
           raise TypeError, "Expected Stoplight::DataStore::Redis, got #{data_store_config.class}"
         end
+      end
+
+      def build_backend
+        case data_store_config
+        in DataStore::Memory
+          Memory::Backend.new(clock:, config:)
+        in DataStore::Redis
+          Redis::Backend.new(
+            redis:, scripting: storage_scripting, key_space:, clock:, config:, error_notifier:,
+            failover_light: create_circuit_breaker("redis")
+          )
+        end
+      end
+
+      def create_circuit_breaker(name)
+        failover_system.light(name)
       end
     end
   end
