--- sig/_private/stoplight/system/light_factory.rbs
@@ -1,23 +1,0 @@
-module Stoplight
-  module Wiring
-    class System
-      class LightFactory < Wiring::LightFactory
-        attr_reader system_name: String
-
-        @failover_system: _System?
-        @storage_set: StorageSet
-        @key_space: Infrastructure::Redis::Storage::KeySpace
-
-        def initialize: (config: Domain::Config, system_name: String, failover_system: _System?) -> void
-
-        def key_space: -> Infrastructure::Redis::Storage::KeySpace
-
-        private def redis: -> redis
-        private def storage_scripting: -> Infrastructure::Redis::Storage::Scripting
-        private def failover_system: -> _System
-        private def storage_set: -> StorageSet
-        private def build_backend: -> DataStoreBackend
-      end
-    end
-  end
-end
