+++ sig/_private/stoplight/wiring/light_factory.rbs
@@ -23,15 +23,21 @@ module Stoplight
       attr_reader traffic_recovery: Domain::_TrafficRecovery
       attr_reader traffic_control: Domain::_TrafficControl
       attr_reader config: Domain::Config
+      attr_reader system_name: String
 
       @notifiers: Array[state_transition_notifier]
       @name: String
       @cool_off_time: duration
       @error_tracking_policy: Domain::ErrorTrackingPolicy
       @wrapped_notifiers: Array[state_transition_notifier]?
+      @failover_system: _System?
+      @storage_set: StorageSet
+      @key_space: Infrastructure::Redis::Storage::KeySpace
 
-      def initialize: (config: Domain::Config) -> void
+      def initialize: (config: Domain::Config, system_name: String, failover_system: _System?) -> void
 
+      def storage_set: -> StorageSet
+      def key_space: -> Infrastructure::Redis::Storage::KeySpace
       def state_store: -> Domain::_StateStore
       def notifiers: -> Array[state_transition_notifier]
 
@@ -45,6 +51,9 @@ module Stoplight
       def yellow_run_strategy: -> Domain::Strategies::YellowRunStrategy
       def red_run_strategy: -> Domain::Strategies::RedRunStrategy
       def redis: -> redis
+      def storage_scripting: -> Infrastructure::Redis::Storage::Scripting
+      def failover_system: -> _System
+      def build_backend: -> DataStoreBackend
 
       # Builds a Light instance with the current configuration.
       #
